Run Chromatic on CI.
This does not run on pull requests. Developers can enable chromatic on
their own forks but we may not be able to require those checks to pass
on PR's. Therefore we don't fail the check if changes are detected, and
will run review process on the develop and release branches but not on
individual PR's.